Move code out of lib.rs.
